chore: move person utils under utils
