+++ src/events/death/forms/correctionForm/requester.ts
@@ -17,8 +17,11 @@ import {
   FieldType
 } from '@opencrvs/toolkit/events'
 import { InformantType, InformantTypeKey } from '../pages/informant'
-import { informantMessageDescriptors } from '@countryconfig/events/utils'
-import { IdType, idTypeOptions } from '@countryconfig/events/person'
+import {
+  IdType,
+  idTypeOptions,
+  informantMessageDescriptors
+} from '@countryconfig/events/utils'
 import { nationalIdValidator } from '@countryconfig/events/birth/validators'
 
 const commonConfigs = {
