+++ src/events/death/forms/pages/deceased.ts
@@ -22,6 +22,9 @@ import {
 } from '@opencrvs/toolkit/events'
 import { not, never } from '@opencrvs/toolkit/conditionals'
 import {
+  IdType,
+  idTypeOptions,
+  maritalStatusOptions,
   createSelectOptions,
   emptyMessage,
   defaultStreetAddressConfiguration,
@@ -32,11 +35,6 @@ import {
   invalidNameValidator,
   nationalIdValidator
 } from '@countryconfig/events/birth/validators'
-import {
-  IdType,
-  idTypeOptions,
-  maritalStatusOptions
-} from '@countryconfig/events/person'
 import {
   connectToMOSIPIdReader,
   connectToMOSIPVerificationStatus,
