+++ src/events/utils/index.ts
@@ -3,3 +3,4 @@ export * from './select-options'
 export * from './street-address-configuration'
 export * from './messages'
 export * from './constants'
+export * from './person'
